+++ modules/gui/macosx/applescript.m
@@ -207,7 +207,7 @@
     if(!p_input)
         return i_delay;
 
-    i_delay = var_GetInteger(p_input, "audio-delay");
+    i_delay = (int)var_GetInteger(p_input, "audio-delay");
     vlc_object_release(p_input);
 
     return (i_delay / 1000);
@@ -224,7 +224,7 @@
 
 - (int) currentTime {
     input_thread_t * p_input = pl_CurrentInput(getIntf());
-    int64_t i_currentTime = -1;
+    int i_currentTime = -1;
 
     if (!p_input)
         return i_currentTime;
@@ -248,7 +248,7 @@
     }
 }
 
-- (int) durationOfCurrentItem {
+- (NSInteger) durationOfCurrentItem {
     return [[VLCCoreInteraction sharedInstance] durationOfCurrentPlaylistItem];
 }
 
@@ -266,7 +266,7 @@
     if (!p_input_thread)
         return NO;
 
-    int i_current_title = var_GetInteger(p_input_thread, "title");
+    int i_current_title = (int)var_GetInteger(p_input_thread, "title");
 
     input_title_t **p_input_title;
     int count;
